Reorganize components into dedicated directories
